Updating the version to 3.1.29.2.P for now.
Creating a new constant variable called EE_LANGUAGES_LOC. This will allow us to point to the languages folder as needed.

Creating a new directory in the wp-content/uploads/espresso/languages/ when the plugin is installed/activated. There is also a blank index.php file created at the same time.

[touch:877]