Creating a global variable called $ee_kses_allowed to store all of the allowed tags.
Adding wp_kses() to several areas of data storage in the plugin such as categories, events, and venues.

Still need to do a complete security lockdown in some places, such as categories, venues, coupon codes, etc. I will create a new ticket for that, but I think this covers the most notable areas that need to be secured.

[touch:2079]