fixes &pound; UTF character issue on invoices
Changed the character encoding for the currency symbol from UTF-8 to ISO-8859-15. The issue was a result of using UTF-8 characters on a page (in this case, the invoice itself) that was using ISO-8859-15. The problem being, of course, that we'd both *like* to use UTF-8 better anyway, and we're converting UTF-8 to ISO-8859-15 later. This just encodes it correctly to begin with.

Also changed "Euro: " to &euro; ISO code rather than Hugo's hokey euro labelling.
[needs testing: 2010]