add .gitattributes file for normalizing line endings. [touch: 2288]
All developers. when pulling this you may all of a sudden see your git repo tell you to commit all your files. DON'T, instead follow the instructions at the bottom of this page here -> https://help.github.com/articles/dealing-with-line-endings OR if you are uncomfortable with the command line, delete your existing copy of the repo and then re clone the repo.