hacks i18n for payment overview page
this is a hack to localize the output of the payment status. rather than addressing the `$payment_status` variable in every. single. gateway., I'm localizing the output with a switch that checks for the payment status and adds a textdomain to it, then spit out the variable. This will require updating the pot file and the language files, which I will do after i look at #2292. [touch:2293]