@brent Not sure if this was intentional or not, but I noticed that the $today_filter variable was being checked on the category joins on line 131, that didn't look right, so I changed it to match what is on line 224. If it was supposed to be the other way, please let me know.
Removed 'get_number_of_attendees_reg_limit' function from this page, because it was calling 3 or 4 additional queries to get the number of attendees and reg limit for each record that was returned.

Removed the 'event_espresso_get_time'function from this page because is was an additional query to get the start time for each record that was returned.

[touch:2514]

Signed-off-by: Seth <seth@eventespresso.com>