do validation on attendee first name, last name, and email fields when adding a new attendee in the backend to make sure that required fields are present. Return error for any fields that are not present (and don't save attendee) [touch: 2923]
Please Note:

For `add_attendees_to_db()` it appears that the function is used in the frontend as well.  So I put my work behind a conditional check for whether its being executed in the admin.  I'm assuming frontend has its own validation being done so I didn't want to potentially mess anything up with registrations of attendees on the frontend.