register and enqueue validation scripts for the shortcodes that need to load them [touch: 3964]
the validation scripts stopped loading for the registration page shortcodes. likely due to some recent changes in how we're now properly enqueueing scripts. I made a new function that registers only the scripts for form validation and added new wp_enqueue_script calls for the shortcodes that need form validation. I left the other function in place that loads up the validation and relCopy scripts because that still works as expected for other parts of the plugin.

I also removed the $load_espresso_scripts global from the shortcodes that display attendee lists and event lists since it's not working and not needed. [touch: 3694]