wrap espresso_check_ssl() function in an if !function exists check [touch: 5452]
I tried using a few checks like ini_get('allow_url_fopen') and this still blew up this one customer's admin. Also tried a check that Darren suggested in chat that checked for SSL with curl. It worked on Sidney's server, but also blew up this customer's admin.

For now, the simple solution (but not automatic) looks like just add the ability to override this little function.