Add another placeholder [touch: 5720]
Unless limit does get specified in the short code, adding a placeholder for $limit alone doesn't prevent the doing_it_wrong notice from being thrown because we set limit to 0 by default, so that part of the query isn't run. We always check for is_active, so that should suffice.