Add a new do action hook in process-registration/payment_page.php for general use. Change the add more attendees link function to use this new hook [touch: 5916]
This will ensure the add more attendees link will appear after the session is completed. For paid events: You'll have to either complete payment or in the case of an offline payment method click through to complete the registration. For free events: the other hook change makes it so it displays on the thank you page when it's a free event registration.